requirements.txt
pyodbc
pandas
matplotlib
streamlit
python-dotenv
requests